Extract common test code into StreamToTableJoinScalaIntegrationTestBase
